crawler/pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crawler"
version = "0.1.0"
dependencies = [
    "accelerate>=0.23.0",
    "bitsandbytes>=0.41.1",
    "datasets>=2.14.4",
    "einops>=0.6.1",
    "evaluate>=0.4.0",
    "mosaicml[wandb]>=0.16.3",
    "polars>=0.19.0",
    "python-dotenv>=1.00",
    "sentence-transformers>=2.2.2",
    "torch==2.1.2",
    "torchvision==0.16.2",
    "torchmetrics>=1.0.0",
    "tqdm>=4.65.0",
    "transformers>=4.36.0",
    "wandb>=0.15.10",
    "zstandard>=0.21.0",
    "hf-transfer>=0.1",
    "zstandard>=0.22.0",
    "wheel",
    "statsmodels",
    "psycopg[binary]>=3.1.18",
    "black>=23.7.0",
    "ipykernel>=6.25.1",
    "ruff>=0.2.0",
    "loguru==0.7.2",
    "httpx==0.27.0",
    "orjson==3.9.15",
    "pyvespa>=0.39.0",
    "pydantic>=2.6.4"
]

[tool.setuptools]
packages = ["crawler"]


[project.optional-dependencies]
gpu = ["vllm==0.3.3", "flash-attn==2.5.5"]
